+++ gpaw/parameters.py
@@ -253,7 +253,6 @@ class InputParameters(dict):
                 self.dtype = complex
                 
         try:
-            from ase.io.jsonio import decode
             from gpaw.external import dict2potential
             self.external = dict2potential(r['ExternalPotential'])
         except KeyError: